refactor: move queries
